fix: move asides into search chunk
